pyproject.toml
[tool.poetry]
name = "cscan-01-23"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cscan_01"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
psycopg2-binary = "^2.9.5"
python-environ = "^0.4.54"
djangorestframework = "^3.14.0"
selenium = "^4.7.2"
webdriver-manager = "^3.8.5"
bs4 = "^0.0.1"
lxml = "^4.9.2"
nltk = "^3.8.1"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.4"
pymorphy2 = "^0.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
